update docs syntax highlighting, remove unsused code
